feat: update and move middleware.ts
